requirements.txt
blinker==1.7.0
certifi>=2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.21
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.1
scapy==2.5.0
pycryptodome==3.20.0
pem==23.1.0
cabarchive==0.2.4